octeontx2-af: Support installing TX MCAM rules
commit 614f9226127ef8b590f4050d349dbc566a1f979c from
git@git.assembla.com:cavium/WindRiver.linux.git

There is no support for installing a transmit
side rule in MCAM currently. This patch adds
support for it. Implementation is similar to
RX where extraction profile is scanned and
location of packet fields are noted and used
later when installing a rule. Direction is also
displayed in debugfs.

~# cat /sys/kernel/debug/octeontx2/npc/mcam_rules
        PF1 VF0
        direction: TX
        mcam entry: 192
        dmac  00:00:00:00:00:00 00:00:00:00:00:00
        action: Drop
        enabled: yes
        counter: 0
        hits: 513

Change-Id: I5747d1cc2afa24fea61f0e7cb7305d056810689d
Signed-off-by: Subbaraya Sundeep <sbhatta@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/13209
Reviewed-by: Sunil Kovvuri Goutham <Sunil.Goutham@cavium.com>
Tested-by: Sunil Kovvuri Goutham <Sunil.Goutham@cavium.com>
[RQ: Replace NPC_LT_LB_STAG with NPC_LT_LB_STAG_QINQ.]
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>