octeontx2-bphy-netdev: Add timeout to ptp transmit timestamp polling
commit 5bf0f82a58f41503153dfac91beda6c9e0d7c358 from
git@git.assembla.com:cavium/WindRiver.linux.git

When there is any error due to which PTP transmit timestamp
is not updated by hardware, the code indefinitely checks for
timestamp valid bit to be set which will be never set. This
stalls the PTP queue processing in the driver. This patch adds
a timeout when checking the valid bit. Also added ethtool stats
to display the stats of ptp timestamp polling timeout.

Change-Id: I6504c4866bb5eca8283f1db04df84c91f850e6ad
Signed-off-by: Naveen Mamindlapalli <naveenm@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/45132
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <Sunil.Goutham@cavium.com>
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>